+++ tests/lib/chat/test_completions_streaming.py
@@ -63,6 +63,7 @@ def test_parse_nothing(client: OpenAI, respx_mock: MockRouter, monkeypatch: pyte
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[NoneType](
+            annotations=None,
             audio=None,
             content="I'm unable to provide real-time weather updates. To get the current weather in San Francisco, I 
 recommend checking a reliable weather website or a weather app.",
@@ -141,6 +142,7 @@ ParsedChatCompletion[Location](
             index=0,
             logprobs=None,
             message=ParsedChatCompletionMessage[Location](
+                annotations=None,
                 audio=None,
                 content='{"city":"San Francisco","temperature":61,"units":"f"}',
                 function_call=None,
@@ -318,6 +320,7 @@ def test_parse_pydantic_model_multiple_choices(
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[Location](
+            annotations=None,
             audio=None,
             content='{"city":"San Francisco","temperature":65,"units":"f"}',
             function_call=None,
@@ -332,6 +335,7 @@ def test_parse_pydantic_model_multiple_choices(
         index=1,
         logprobs=None,
         message=ParsedChatCompletionMessage[Location](
+            annotations=None,
             audio=None,
             content='{"city":"San Francisco","temperature":61,"units":"f"}',
             function_call=None,
@@ -346,6 +350,7 @@ def test_parse_pydantic_model_multiple_choices(
         index=2,
         logprobs=None,
         message=ParsedChatCompletionMessage[Location](
+            annotations=None,
             audio=None,
             content='{"city":"San Francisco","temperature":59,"units":"f"}',
             function_call=None,
@@ -421,6 +426,7 @@ RefusalDoneEvent(refusal="I'm sorry, I can't assist with that request.", type='r
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[Location](
+            annotations=None,
             audio=None,
             content=None,
             function_call=None,
@@ -495,6 +501,7 @@ def test_content_logprobs_events(client: OpenAI, respx_mock: MockRouter, monkeyp
             refusal=None
         ),
         message=ParsedChatCompletionMessage[NoneType](
+            annotations=None,
             audio=None,
             content='Foo!',
             function_call=None,
@@ -606,6 +613,7 @@ def test_refusal_logprobs_events(client: OpenAI, respx_mock: MockRouter, monkeyp
             ]
         ),
         message=ParsedChatCompletionMessage[Location](
+            annotations=None,
             audio=None,
             content=None,
             function_call=None,
@@ -652,6 +660,7 @@ def test_parse_pydantic_tool(client: OpenAI, respx_mock: MockRouter, monkeypatch
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[object](
+            annotations=None,
             audio=None,
             content=None,
             function_call=None,
@@ -684,6 +693,7 @@ def test_parse_pydantic_tool(client: OpenAI, respx_mock: MockRouter, monkeypatch
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[NoneType](
+            annotations=None,
             audio=None,
             content=None,
             function_call=None,
@@ -755,6 +765,7 @@ def test_parse_multiple_pydantic_tools(client: OpenAI, respx_mock: MockRouter, m
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[object](
+            annotations=None,
             audio=None,
             content=None,
             function_call=None,
@@ -863,6 +874,7 @@ def test_parse_strict_tools(client: OpenAI, respx_mock: MockRouter, monkeypatch:
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[object](
+            annotations=None,
             audio=None,
             content=None,
             function_call=None,
@@ -914,6 +926,7 @@ def test_non_pydantic_response_format(client: OpenAI, respx_mock: MockRouter, mo
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[NoneType](
+            annotations=None,
             audio=None,
             content='\\n  {\\n    "location": "San Francisco, CA",\\n    "weather": {\\n      "temperature": "18°C",\\n      
 "condition": "Partly Cloudy",\\n      "humidity": "72%",\\n      "windSpeed": "15 km/h",\\n      "windDirection": "NW"\\n   
@@ -974,6 +987,7 @@ FunctionToolCallArgumentsDoneEvent(
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[NoneType](
+            annotations=None,
             audio=None,
             content=None,
             function_call=None,
@@ -1033,6 +1047,7 @@ def test_chat_completion_state_helper(client: OpenAI, respx_mock: MockRouter, mo
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[NoneType](
+            annotations=None,
             audio=None,
             content="I'm unable to provide real-time weather updates. To get the current weather in San Francisco, I 
 recommend checking a reliable weather website or a weather app.",